% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_decisions_df}
\alias{build_decisions_df}
\title{Build Decisions Data.Frame}
\usage{
build_decisions_df(
  decision_lst = c("Low", "Medium", "High"),
  rules_lst = list(downloads_1yr = list(cond = list(High = ~is.na(.x), High = ~.x <
    120000, Medium = ~dplyr::between(.x, 120000, 240000), Low = ~.x > 240000), type =
    "primary", accept_cats = c("Low"), min_value = 40000), reverse_dependencies =
    list(cond = list(High = ~is.na(.x), High = ~.x < 2, Medium = ~dplyr::between(.x, 2,
    7), Low = ~.x > 7), type = "exception", accept_cats = c("Low", "Medium")),
    dependencies = list(cond = list(High = ~is.na(.x), High = ~.x > 8, Medium =
    ~dplyr::between(.x, 4, 8), Low = ~.x < 
     4), type = "exception", accept_cats =
    c("Low", "Medium")), news_current = list(cond = list(High = ~is.na(.x), High = ~.x !=
    1, Low = ~.x == 1), type = "exception", accept_cats = c("Low")), has_vignettes =
    list(cond = list(High = ~is.na(.x), High = ~.x < 1, Medium = ~.x == 1, Low = ~.x >
    1), type = "exception", accept_cats = c("Low")), has_source_control = list(cond =
    list(High = ~is.na(.x), High = ~.x == 0, Low = ~.x > 0), type = "exception",
    accept_cats = c("Low")), has_website = list(cond = list(High = ~is.na(.x), 
     High
    = ~.x == 0, Low = ~.x > 0), type = "exception", accept_cats = c("Low")))
)
}
\arguments{
\item{decision_lst}{A character vector of decision categories, ordered from
highest risk to lowest risk.}

\item{rules_lst}{A named list of lists, where each sub-list contains:
\itemize{
\item cond: A named list of formulas, where the names correspond to decision
categories in \code{decision_lst}, and the formulas define the conditions for
each category.
\item type: A character string indicating whether the metric is "primary" or
"exception".
\item accept_cats: A character vector of decision categories (& their conditions)
that are acceptable for this metric.
\item min_value: (Optional) A numeric value indicating the minimum acceptable
value for the primary metric(s).
}}
}
\value{
A data.frame with columns:
\itemize{
\item metric: The name of the metric.
\item decision: The decision category.
\item condition: The condition as a string.
\item lower_limit: The lower limit of the condition.
\item upper_limit: The upper limit of the condition.
\item metric_type: The type of the metric ("Primary" or "Exception").
\item accept_condition: A logical indicating whether the condition is acceptable.
}
}
\description{
Helper function that creates the minimally necessary data.frame for
val_filter() that includes columns "metric", "decision", "condition",
"metric_type", and "accept_condition".
}
\examples{
build_decisions_df()

}
