% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decision_to_id}
\alias{decision_to_id}
\title{Decision to ID}
\usage{
decision_to_id(decision_id_df, rev = FALSE, dec)
}
\arguments{
\item{decision_id_df}{A data.frame with columns "decision" and "decision_id"}

\item{rev}{Logical. If TRUE, convert decision_id to decision category, else convert decision category to decision_id}

\item{dec}{A character string (if rev = FALSE) or numeric/integer (if rev = TRUE) indicating the decision category or decision_id to convert}
}
\description{
Helper function to convert decision category to decision_id or vice versa
}
\examples{
dec_id_df <- unique(build_decisions_df()[c("decision", "decision_id")])
decision_to_id(dec_id_df, FALSE, "High")

}
\keyword{internal}
