% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decision_to_id_v}
\alias{decision_to_id_v}
\title{Vectorized Decision to ID}
\usage{
decision_to_id_v(decision_id_df, rev = FALSE, dec)
}
\arguments{
\item{decision_id_df}{A data.frame with columns "decision" and "decision_id"}

\item{rev}{Logical. If TRUE, convert decision_id to decision category, else convert decision category to decision_id}

\item{dec}{A character vector (if rev = FALSE) or numeric/integer vector (if rev = TRUE) indicating the decision categories or decision_ids to convert}
}
\description{
Vectorized version of decision_to_id()
}
\examples{
\dontrun{
dec_id_df <- unique(build_decisions_df("remote_reduce")[c("decision", "decision_id")])
decision_to_id_v(dec_id_df, FALSE, c("High", "Medium", "Low"))
}

}
\keyword{internal}
