% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%|e|\%}
\alias{\%|e|\%}
\title{Empty string Coalesce}
\usage{
x \%|e|\% y
}
\arguments{
\item{x, y}{Two elements to test, one potentially empty}
}
\description{
If x has zero characters, return y; otherwise return x.
}
\examples{
"" \%|e|\% 1
}
