% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem_watchdog.R
\name{mem_watchdog_cols}
\alias{mem_watchdog_cols}
\title{Column schema for \code{mem_watchdog.tsv}}
\usage{
mem_watchdog_cols()
}
\description{
Column schema for \code{mem_watchdog.tsv}
}
\keyword{internal}
