% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rip_cats}
\alias{rip_cats}
\title{Generate Decision Category Assignments}
\usage{
rip_cats(met_dec_df, pkgs_df, else_cat)
}
\arguments{
\item{met_dec_df}{A data.frame of decisions, typically created by
build_decisions_df()}

\item{pkgs_df}{A data.frame of package metrics, typically created by
available.packages()}

\item{else_cat}{A character string indicating the default category to assign
when none of the conditions are met.}
}
\description{
A helper function that applies the dplyr::case_when() statements from the
decisions data.frame to the package metrics data.frame, creating new columns
with the decision categories for each metric.
}
\examples{
# rip_cats(
#   build_decisions_df() |> dplyr::mutate(derived_col = metric),
#   available.packages() |> 
#     as.data.frame() |> 
#     dplyr::select(
#       pkg, 
#       downloads_1yr = dplyr::starts_with("downloads_1yr")
#     ),
#   "High"
# )

}
\keyword{internal}
