% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rip_cats_by_pkg}
\alias{rip_cats_by_pkg}
\title{Generate Decision Category Assignments by Package}
\usage{
rip_cats_by_pkg(
  label = "primary",
  repo_name = "CRAN",
  dec_df = dplyr::mutate(build_decisions_df(rule_type = "decide"), derived_col = metric),
  pkgs_df = NULL,
  decisions = pull_config(val = "decisions_lst", rule_type = "default"),
  else_cat = NULL
)
}
\arguments{
\item{label}{A character string indicating the metric type to process}

\item{repo_name}{A character string indicating the repository name}

\item{dec_df}{A data.frame of decisions, typically created by
build_decisions_df()}

\item{pkgs_df}{A data.frame of package metrics, typically created by
available.packages() merged with riskmetric assessments}

\item{decisions}{A character vector of decision categories, typically
obtained from pull_config()}

\item{else_cat}{A character string indicating the default category to assign
when none of the conditions are met.}
}
\description{
A helper function that applies the dplyr::case_when() statements from the
decisions data.frame to the package metrics data.frame, creating new columns
with the decision categories for each metric.
}
\keyword{internal}
