% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_join_cats}
\alias{split_join_cats}
\title{Generate Decision Category Assignments, splitting out downloads conditions
for only CRAN pkgs}
\usage{
split_join_cats(
  pkgs_data = NULL,
  dec_df = build_decisions_df(rule_type = "remote_reduce"),
  decisions = unique(dec_df$decision),
  else_cat = decisions[length(decisions)],
  label = "Primary"
)
}
\arguments{
\item{pkgs_data}{A data.frame of package metrics, typically created by
available.packages()}

\item{dec_df}{A data.frame of decisions, typically created by
build_decisions_df()}

\item{decisions}{A character vector of decision categories, ordered from
highest risk to lowest risk.}

\item{else_cat}{A character string indicating the default category to assign
when none of the conditions are met.}

\item{label}{A character string indicating the label to use in messages.}
}
\value{
A data.frame of package metrics with decision categories assigned.
}
\description{
A helper function that splits the package metrics data.frame into CRAN and
non-CRAN packages, applies the rip_cats() function to each subset based on
the applicable metrics, and then rejoins the two data.frames.
}
\keyword{internal}
