% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_recording_df}
\alias{strip_recording_df}
\title{Strip Recording (Data Frame) - Deprecated}
\usage{
strip_recording_df(assessment)
}
\arguments{
\item{assessment}{A list() object, typically the output of pkg_assess()}
}
\description{
Remove .recording attribute from all elements of the assessment,
whilst maintaining classes
}
\examples{
# riskmetric::pkg_ref("zoo", source = "pkg_cran_remote") |>
#   dplyr::as_tibble() |>
#   riskmetric::pkg_assess() |>
#   strip_recording_df()

}
\keyword{internal}
