% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_the_limit}
\alias{to_the_limit}
\title{To the Limit}
\usage{
to_the_limit(condition, low = TRUE)
}
\arguments{
\item{condition}{A character vector of conditions (as strings)}

\item{low}{Logical. If TRUE, extract the lower limit, else the upper limit}
}
\description{
Helper function to parse lower & upper limit out of the condition string
}
\examples{
to_the_limit("~ is.na(.x)", low = FALSE)
to_the_limit("~ .x < 120000", low = FALSE)
to_the_limit("~ dplyr::between(.x, 120000, 240000)", low = FALSE)
to_the_limit("~ .x > 240000", low = FALSE)

}
\keyword{internal}
