% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_opt_repos}
\alias{update_opt_repos}
\title{Update Repos Option}
\usage{
update_opt_repos(val_date = Sys.Date(), opt_repos = getOption("repos"))
}
\arguments{
\item{val_date}{A Date object indicating the validation date to use in the
CRAN repo URL.}

\item{opt_repos}{A named character vector of repositories, typically obtained
from getOption("repos").}
}
\value{
A named character vector of repositories with the updated "CRAN"
repo.
}
\description{
Helper function to update the "CRAN" repo in options("repos") to use the
specified validation date.
}
