% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_decision.R
\name{val_categorize}
\alias{val_categorize}
\title{Categorize Decisions for A Corpus of Packages}
\usage{
val_categorize(
  source = "riskscore",
  decisions = c("Low", "Medium", "High"),
  else_cat = "High",
  decisions_df = build_decisions_df(rule_type = "remote_reduce"),
  verbose = NULL
)
}
\arguments{
\item{source}{character, either "riskscore" (default), "PACKAGES", or a
data.frame.}

\item{decisions}{character vector, the risk categories to use.}

\item{else_cat}{character, the default risk category if no conditions are met.}

\item{decisions_df}{data.frame, the output of build_decisions_df()}

\item{verbose}{Console verbosity control. One of \code{"quiet"},
\code{"minimal"}, \code{"normal"} (default), or \code{"verbose"}. See
the \code{val.pipeline} verbosity docs for tier definitions.}
}
\description{
First whack at attempting to filter packages based on org-level criterion
used to set thresholds (and later update final decision (if not already 'high
risk')) AND filter packages before running val_build(). Note: If PACKAGES
file had assessments, we'd be using that (paired with \{val.filter\}), but
instead, we use \code{riskscore::assessed_latest} / \code{riskscore::scored_latest}
(compiled from the \code{latest} branch of pharmar/riskscore).
}
