% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_decision.R
\name{val_decision}
\alias{val_decision}
\title{Assign a decision}
\usage{
val_decision(
  pkg = NULL,
  source_df = NULL,
  excl_metrics = NULL,
  decisions = c("Low", "Medium", "High"),
  else_cat = "High",
  avail_pkgs = as.data.frame(available.packages()),
  decisions_df = build_decisions_df(rule_type = "decide")
)
}
\arguments{
\item{pkg}{character, the package name to assess}

\item{excl_metrics}{character vector, allow users to subset metrics as needed
without changing the config}

\item{decisions}{character vector, the risk categories to use.}

\item{else_cat}{character, the default risk category if no conditions are
met.}

\item{avail_pkgs}{data.frame, the output of available.packages()}

\item{decisions_df}{data.frame, the output of build_decisions_df()}

\item{source}{data.frame, comprised of \code{riskmetric} assessments & scores,
usually the output of \code{workable_assessments()}.}
}
\description{
Categorize packages based on org-level criterion / thresholds at the
package-level. This function applies the decision rules to a single package
and returns the risk category along with the relevant metrics used to make
the decision. This function assumes that the package metrics have already
been computed and are available in the \code{source} parameter.#'
}
