% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_filter.R
\name{val_filter}
\alias{val_filter}
\title{Filter Packages}
\usage{
val_filter(
  pre = TRUE,
  source = "riskscore",
  avail_pkgs = as.data.frame(available.packages())
)
}
\arguments{
\item{pre}{logical, whether or not we are using pre or post filtering logic.
May remove this later.}

\item{source}{character, either "riskscore" (default), "PACKAGES", or a
data.frame.}

\item{avail_pkgs}{data.frame, the output of available.packages() as a data.frame}
}
\description{
First whack at attempting to filter packages based on org-level criterion
used to set thresholds (and later update final decision (if not already 'high
risk')) AND filter packages before running val_build(). Note: If PACKAGES
file had assessments, we'd be using that (paired with {val.filter}), but
instead, we're going to use riskscore::cran_assessed_20250812 for the time
being
}
