% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{workable_assessments}
\alias{workable_assessments}
\title{Prepare Workable Assessments DF}
\usage{
workable_assessments(
  pkg = NULL,
  ver = NULL,
  val_date = Sys.Date(),
  metric_pkg = c("riskmetric", "val.meter", "risk.assessr"),
  source = NULL,
  src_ref = c("source", "remote")
)
}
\arguments{
\item{pkg}{The name of the package}

\item{ver}{The version of the package}

\item{val_date}{The validation date}

\item{metric_pkg}{The package used to generate the package assessment.
Options are "riskmetric", "val.meter", or "risk.assessr"}

\item{source}{The source object containing the package assessment.
Typically the output of pkg_assess() & pkg_score() from the \code{riskmetric}.}

\item{src_ref}{A character string indicating the source reference,
either "source" (local pkg_source) or "remote" (CRAN/pkg_remote)}
}
\value{
A data.frame/tibble with workable fields for a val_decision() call
}
\description{
Helper function to extract workable fields from a package assessment
for use in val_pkg() where val_decision() is called.
}
\keyword{internal}
