% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_decisions_df}
\alias{build_decisions_df}
\title{Build Decisions Data.Frame}
\usage{
build_decisions_df(
  rule_type = c("default", "remote_reduce", "decide")[1],
  rule_lst = NULL
)
}
\arguments{
\item{rule_type}{A character string indicating whether the decision
categories are used to "categorize" risk levels (e.g., "Low", "Medium",
"High") via val_categorize() or val_decision(). The difference being that
"categorize" is going to be used to filter the initial list of packages
using a remote pkg_ref() assessment, whereas "decision" is used to filter
the final list of packages after a 'pkg_source' pkg_ref() assessment is
produced locally on the GxP system of interest.}

\item{rule_lst}{A named list of lists, where each sub-list contains:
\itemize{
\item cond: A named list of formulas, where the names correspond to decision
categories in \code{decision_lst}, and the formulas define the conditions for
each category.
\item type: A character string indicating whether the metric is "primary" or
"exception".
\item accept_cats: A character vector of decision categories (& their conditions)
that are acceptable for this metric.
\item min_value: (Optional) A numeric value indicating the minimum acceptable
value for the primary metric(s).
}}

\item{decision_lst}{A character vector of decision categories, ordered from
highest risk to lowest risk.opt_repos}
}
\value{
A data.frame with columns:
\itemize{
\item metric: The name of the metric.
\item decision: The decision category.
\item condition: The condition as a string.
\item lower_limit: The lower limit of the condition.
\item upper_limit: The upper limit of the condition.
\item metric_type: The type of the metric ("Primary" or "Exception").
\item accept_condition: A logical indicating whether the condition is acceptable.
}
}
\description{
Helper function that creates the minimally necessary data.frame for
val_filter() that includes columns "metric", "decision", "condition",
"metric_type", and "accept_condition".
}
\examples{
build_decisions_df()

}
