% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_repo_origin}
\alias{get_repo_origin}
\title{Get Repo Origin}
\usage{
get_repo_origin(repo_src = NULL, pkg_name = NULL, names_only = FALSE)
}
\arguments{
\item{repo_src}{character, containing a repository URL for a single pkg}

\item{pkg_name}{The name of the package (just for warning messages)}
}
\value{
A character string indicating the name of the repository
}
\description{
Helper function to determine which repo a package came from
}
\examples{
get_repo_origin(repo_src = "https://packagemanager.posit.co/cran/2024-06-01", "fake")

}
\keyword{internal}
