% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_config}
\alias{pull_config}
\title{Pull Config}
\usage{
pull_config(
  val = NULL,
  rule_type = c("default", "remote_reduce", "decide")[1],
  config_path = system.file("config.yml", package = "val.pipeline")
)
}
\arguments{
\item{val}{Name of value (NULL to read all values)}

\item{rule_type}{A character string indicating whether the decision
categories are used to "categorize" risk levels (e.g., "Low", "Medium",
"High") via val_categorize() or val_decision(). The difference being that
"categorize" is going to be used to filter the initial list of packages
using a remote pkg_ref() assessment, whereas "decision" is used to filter
the final list of packages after a 'pkg_source' pkg_ref() assessment is
produced locally on the GxP system of interest.}

\item{config_path}{A character string indicating the path to the config.yml}
}
\value{
A named list of lists
}
\description{
Pull in relevant rules defined in confg.yml. Also verifies that the
conditions only reference decision categories that are allowed in
\code{decision_lst}.
}
\examples{
pull_config(rule_type = "remote_reduce")
pull_config("decide_github", rule_type = "remote_reduce")

}
