% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_recording}
\alias{strip_recording}
\title{Strip Recording}
\usage{
strip_recording(assessment)
}
\description{
Remove .recording attribute from all elements of the assessment,
whilst maintaining classes
}
\keyword{internal}
