% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_build.R
\name{val_build}
\alias{val_build}
\title{Build Validation}
\usage{
val_build(
  r_ver = getRversion(),
  pkg_names = NULL,
  deps = c("depends", "suggests"),
  out = "riskassessment",
  opt_repos = c(val_build_repo = "https://cran.r-project.org")
)
}
\description{
Build Validation
}
