% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_build.R
\name{val_build}
\alias{val_build}
\title{Validation: Build an Assessment Co-hort}
\usage{
val_build(
  pkg_names = NULL,
  ref = c("source", "remote"),
  metric_pkg = c("riskmetric", "val.meter", "risk.assessr"),
  deps = c("depends", "suggests")[1],
  deps_recursive = TRUE,
  val_date = Sys.Date(),
  out = "riskassessment",
  replace = FALSE,
  opt_repos = c(CRAN = "https://packagemanager.posit.co/cran/latest", BioC =
    "https://bioconductor.org/packages/3.21/bioc")
)
}
\arguments{
\item{pkg_names}{Character vector of package names to assess. If NULL
(default), all packages available from the specified repository will be
assessed.}

\item{ref}{Character string indicating the source of the packages to assess.
Either "source" (default) for source packages, or "remote" for packages
from remote repositories like CRAN/Bioconductor.}

\item{metric_pkg}{Character string specifying the risk assessment package to
use. Either "riskmetric" (default) & or "val.meter" (not implemented yet).}

\item{deps}{Character vector specifying which types of dependencies to
include in the assessment. Options are "depends", "suggests", or both
(default). If NULL, only the specified packages from 'pkg_names' will be
assessed without their dependencies.}

\item{deps_recursive}{Logical indicating whether to include dependencies
recursively. Default is TRUE.}

\item{val_date}{Date object or character string representing the date of the
validation build. Default is the current date (Sys.Date()).}

\item{out}{Character string specifying the output directory for the
validation build. Default is "riskassessment" in the current working
directory.}

\item{replace}{Logical indicating whether to replace existing assessments for
packages that have already been assessed. Default is FALSE.}

\item{opt_repos}{Named character vector specifying the repository options for
package installation. Default is CRAN.}
}
\value{
A list containing:
\itemize{
\item val_dir: The directory where the validation build results are stored.
\item pkgs_df: A data frame summarizing the risk assessment results for all packages assessed,
including their dependencies and final risk decisions.
}
}
\description{
Build a risk assessment validation for a set of R packages from various
sources (CRAN / Bioconductor / GitHub), with the ability to include
(optionally recursive) dependencies and suggests, and save the results in a
structured directory. The cherry on top is that this build will use logic
from val_decision() to not only apply risk decisions too all packages
assessed, but goes back around and will re-categorize decisions based on
whether any dependencies were categorized as "High Risk" / "Rejected". It
also is intelligent enough to sort the list of packages to run those with the
most dependencies first, so that if a package fails, it doesn't waste any
time running it's reverse dependence.
After the pipeline applies a decision onto each package using criteria
provided in a config file, it even generates a report detailing specifics of
the assessment as supporting evidence. The end result is a directory
containing the assessment results and reports for each package evaluated.
}
