% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_build.R
\name{val_build}
\alias{val_build}
\title{Build Validation}
\usage{
val_build(
  pkg_names = NULL,
  ref = "source",
  metric_pkg = "riskmetric",
  deps = c("depends", "suggests"),
  val_date = Sys.Date(),
  rerun = FALSE,
  out = "riskassessment",
  opt_repos = c(val_build_repo = "https://cran.r-project.org")
)
}
\description{
Build Validation
}
