% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_decision.R
\name{val_decision}
\alias{val_decision}
\title{Assign a decision}
\usage{
val_decision(
  pkg = NULL,
  source = NULL,
  excl_metrics = NULL,
  decisions = c("Low", "Medium", "High"),
  else_cat = "High",
  decisions_df = build_decisions_df(rule_type = "decide")
)
}
\arguments{
\item{pkg}{character, the package name to assess}

\item{source}{character, either "riskscore" (default), "PACKAGES", or a
data.frame.}

\item{excl_metrics}{character vector, allow users to subset metrics as needed
without changing the config}

\item{decisions}{character vector, the risk categories to use.}

\item{else_cat}{character, the default risk category if no conditions are
met.}

\item{decisions_df}{data.frame, the output of build_decisions_df()}
}
\description{
First whack at attempting to filter packages based on org-level criterion
used to set thresholds (and later update final decision (if not already 'high
risk')) AND filter packages before running val_build(). Note: If PACKAGES
file had assessments, we'd be using that (paired with \{val.filter\}), but
instead, we're going to use riskscore::cran_assessed_20250812 for the time
being
}
