% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_pkg.R
\name{val_pkg}
\alias{val_pkg}
\title{Validation: Assess a Package}
\usage{
val_pkg(
  pkg,
  ver,
  avail_pkgs,
  ref = c("source", "remote"),
  metric_pkg = "riskmetric",
  out_dir,
  val_date = Sys.Date()
)
}
\arguments{
\item{pkg}{Character(1). Name of package to validate.}

\item{ver}{Character(1). Version of package to validate.}

\item{avail_pkgs}{Data frame. Output of \code{available.packages()}.}

\item{ref}{Character(1). Either "source" or "remote". Indicates whether to
download the package source from CRAN (or other repo) or to install the
package directly from the repo.}

\item{metric_pkg}{Character(1). Either "riskmetric", "val.meter", or
"risk.assessr" indicating which package to use for the assessment.}

\item{out_dir}{Character(1). Directory to store outputs.}

\item{val_date}{Date. Date of validation. Default is current date.}
}
\value{
A list containing package metadata, assessment results, and
decisions.
}
\description{
Validation process at the package level. Includes steps to download the
package source (preferred), install the package, assess the package using the
user-specified metric package (only \code{riskmetric} is supported currently),
apply risk decisions, and build a report. Note: to save time during, every
package will be assessed using a "pkg_cran_remote" reference initially to see
if any primary metrics met the "auto-accept" threshold(s), if applicable. If
they did, then the running then computing 'covr_coverage' will be skipped.
}
