% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_pkg.R
\name{val_pkg}
\alias{val_pkg}
\title{Package Validation}
\usage{
val_pkg(
  pkg,
  ver,
  avail_pkgs,
  ref = "source",
  metric_pkg = "riskmetric",
  out_dir,
  val_date = Sys.Date()
)
}
\description{
Validation process at the package level. Includes steps...
}
